scripts/requirements.txt
beautifulsoup4==4.9.3
PyYAML==5.4.1
lxml==4.6.3
html5lib==1.1